% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_function.R
\name{a_function}
\alias{a_function}
\title{A roxygen skeleton added with "Code > Insert Roxygen Skeleton"}
\usage{
a_function()
}
\value{
A string giving a message.
}
\description{
A paragraph giving a description.
}
\details{
Optionally, one or more paragraphs givin details.
}
\examples{
a_function()
}
