% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fair_share_columns.R
\name{add_fair_share_columns}
\alias{add_fair_share_columns}
\title{Add "fair share" columns to a \code{scenario} dataset}
\usage{
add_fair_share_columns(scenario, start_year)
}
\arguments{
\item{scenario}{A scenario dataframe like \link[r2dii.dataraw:scenario_demo]{r2dii.dataraw::scenario_demo}.}

\item{start_year}{The start year of the analysis, crucial for calculating the
correct initial sector total.}
}
\value{
A dataframe with the same groups (if any) and columns as \code{scenario},
and the additional columns:
\itemize{
\item \code{tfsr} - The "Technology Fair Share Ratio", used to calculate scenario
efforts for high-carbon technologies. The portfolio efforts are calculated
as: \eqn{p(t) = p(0) * tfsr}, where \code{p(0)} is the technology-level value at
\code{start_year}.
\item \code{mfsp} - The "Market Fair Share Percentage", used to calculate scenario
efforts for low-carbon technologies. The portfolio efforts are calculated
as \eqn{p(t) = p(0) + (P(0) + mfsp}, where where \code{p(0)} and \code{P(0)} is are
the technology level and sector-level value at \code{start_year}.
}
}
\description{
Add "fair share" columns to a \code{scenario} dataset
}
\examples{
if (requireNamespace("r2dii.dataraw", quietly = TRUE)) {
  scenario <- r2dii.dataraw::scenario_demo
  add_fair_share_columns(scenario, start_year = 2020)
}
}
