% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_loan_weighted_production.R
\name{add_loan_weighted_production}
\alias{add_loan_weighted_production}
\title{Adds column \code{loan_weighted_production}}
\usage{
add_loan_weighted_production(data, use_loan_size_credit_limit = FALSE)
}
\arguments{
\item{data}{A dataset of class "data.frame", like the output of
\code{\link[=join_ald_scenario]{join_ald_scenario()}}.}

\item{use_loan_size_credit_limit}{Logical vector of length 1. \code{FALSE}
defaults to using the column \code{loan_size_outstanding}. Set to \code{TRUE} to use
the column \code{loan_size_credit_limit} instead.}
}
\value{
A dataframe with the same columns as \code{data} and the additional column
\code{loan_weighted_production}, where each row is aggregates information of all
\code{loan_id}s within groups defined by the unique combinations of the columns
\code{sector}, \code{technology}, and \code{year}.
}
\description{
This function computes the production of a portfolio weighted by loan size.
It aggregates the production -- weighted by loan size -- off all loans for
each sector, technology, and year.
}
\examples{
library(r2dii.dataraw)

master <- r2dii.dataraw::loanbook_demo \%>\%
  r2dii.match::match_name(r2dii.dataraw::ald_demo) \%>\%
  r2dii.match::prioritize() \%>\%
  join_ald_scenario(r2dii.dataraw::ald_demo, r2dii.dataraw::scenario_demo)

add_loan_weighted_production(master)

add_loan_weighted_production(master, use_loan_size_credit_limit = TRUE)
}
\seealso{
\code{\link[=join_ald_scenario]{join_ald_scenario()}}.
}
