% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sda_target.R
\name{add_sda_target}
\alias{add_sda_target}
\title{Add portfolio level SDA emissions factor targets}
\usage{
add_sda_target(data, ald, co2_intensity_scenario, use_credit_limit = FALSE)
}
\arguments{
\item{data}{A dataframe like the output of
\code{\link[r2dii.match:prioritize]{r2dii.match::prioritize()}}.}

\item{ald}{An asset level dataframe like \link[r2dii.data:ald_demo]{r2dii.data::ald_demo}.}

\item{co2_intensity_scenario}{A scenario dataframe like
\link[r2dii.analysis:co2_intensity_scenario]{r2dii.analysis::co2_intensity_scenario}.}

\item{use_credit_limit}{Logical vector of length 1. \code{FALSE} defaults to using
the column \code{loan_size_outstanding}. Set to \code{TRUE} to use the column
\code{loan_size_credit_limit} instead.}
}
\value{
A tibble with portfolio-level emission_factor targets.
}
\description{
This function calculates the portfolio-level emissions factor targets, as
calculated using the sectoral-decarbonization approach.
}
\examples{
installed <- requireNamespace("r2dii.data", quietly = TRUE) &&
  requireNamespace("r2dii.match", quietly = TRUE)
if (!installed) stop("Please install r2dii.match and r2dii.data")

library(r2dii.data)
library(r2dii.match)

valid_matches <- match_name(loanbook_demo, ald_demo) \%>\%
  # WARNING: Remember to validate matches (see `?prioritize`)
  prioritize()

valid_matches \%>\%
  add_sda_target(
    ald = ald_demo,
    co2_intensity_scenario = r2dii.analysis::co2_intensity_scenario
  )
}
