% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_weighted_production.R
\name{add_weighted_production}
\alias{add_weighted_production}
\title{Adds column \code{weighted_production}}
\usage{
add_weighted_production(data, use_loan_size_credit_limit = FALSE)
}
\arguments{
\item{data}{A dataframe like the output of \code{\link[=join_ald_scenario]{join_ald_scenario()}}.}

\item{use_loan_size_credit_limit}{Logical. Use \code{TRUE} to use
the column \code{loan_size_credit_limit} instead of the default column
\code{loan_size_outstanding}.}
}
\value{
A dataframe with the same columns as \code{data} and the additional
column \code{weighted_production}, where each row is aggregates information
of all \code{loan_id}s within groups defined by the unique combinations of
the columns \code{sector}, \code{technology}, and \code{year}.
}
\description{
This function computes the weighted production of a portfolio from a matched
dataset (see argument \code{data}). This metric aggregates the weighted production
off all loans for each sector, technology, and year.
}
\examples{
library(r2dii.dataraw)

master <- r2dii.dataraw::loanbook_demo \%>\%
  r2dii.match::match_name(r2dii.dataraw::ald_demo) \%>\%
  r2dii.match::prioritize() \%>\%
  join_ald_scenario(r2dii.dataraw::ald_demo, r2dii.dataraw::scenario_demo)

add_weighted_production(master)

add_weighted_production(master, use_loan_size_credit_limit = TRUE)
}
\seealso{
\code{\link[=join_ald_scenario]{join_ald_scenario()}}.
}
