% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_indicator.R
\name{calculate_temperature_indicator}
\alias{calculate_temperature_indicator}
\title{TODO Add title}
\usage{
calculate_temperature_indicator(
  input_results,
  upper_temp_threshold = 6,
  lower_temp_threshold = 1.5,
  start_year = NULL,
  time_horizon = NULL,
  allocation = NULL,
  production_type = "absolute",
  group_vars = c("Investor.Name", "Portfolio.Name", "Asset.Type")
)
}
\arguments{
\item{scenario_relationships}{}
}
\value{

}
\description{
TODO Add title
}
