% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2_intensity_scenario.R
\docType{data}
\name{co2_intensity_scenario}
\alias{co2_intensity_scenario}
\title{A prepared co2 intensity climate scenario dataset for demonstration}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 22 rows and 7 columns.
}
\usage{
co2_intensity_scenario
}
\description{
Fake co2 intensity climate scenario dataset, prepared for the software PACTA (Paris
Agreement Capital Transition Assessment). It imitates climate scenario data
(e.g. from the International Energy Agency (IEA)) including the change
through time in production across industrial sectors (calculated by
\href{https://2degrees-investing.org/}{2DII}).
}
\examples{
head(co2_intensity_scenario)
}
\keyword{datasets}
