% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Currencies.R
\docType{data}
\name{Currencies}
\alias{Currencies}
\title{Currencies dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 59 rows and 6 columns.}
\usage{
Currencies
}
\description{
TODO: Add a description (1 paragraph).
}
\examples{
Currencies
}
\keyword{datasets}
