% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionary.R
\docType{data}
\name{data_dictionary}
\alias{data_dictionary}
\title{Data Dictionary}
\format{
\subsection{\code{data_dictionary}}{

\describe{
\item{dataset}{Name of the dataset}
\item{column}{Name of the column}
\item{typeof}{Type of the column}
\item{definition}{Definition of the column}
}
}
}
\usage{
data_dictionary
}
\description{
A table of column names and descriptions of data frames used or exported by
the functions in this package.
}
\examples{
data_dictionary
}
\concept{data dictionary}
\keyword{datasets}
