% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fin_data.R
\docType{data}
\name{fin_data}
\alias{fin_data}
\title{Financial data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 174 rows and 4 columns.}
\usage{
fin_data
}
\description{
Financial data
}
\examples{
fin_data
}
\keyword{datasets}
