% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_indicator.R
\name{find_range}
\alias{find_range}
\title{TODO vintented}
\usage{
find_range(input_results, metric_name, range)
}
\arguments{
\item{input_results}{A dataframe with a column with a column that provides
a numeric temperature indicator at varying granularities.}

\item{metric_name}{A character name that specifies a dataframe column with numeric values.}

\item{range}{A list of numeric values that define the range that a
temperature indicators lies within.}
}
\value{

}
\description{
TODO vintented
}
