% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fund_data.R
\docType{data}
\name{fund_data}
\alias{fund_data}
\title{Fund data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 49 rows and 1 columns.}
\usage{
fund_data
}
\description{
Fund data
}
\examples{
fund_data
}
\keyword{datasets}
