% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_current_year}
\alias{get_current_year}
\title{Get the current year from the system time}
\usage{
get_current_year(x)
}
\arguments{
\item{x}{a date-time object}
}
\value{
the years element of x as a decimal number
}
\description{
Date-time must be a POSIXct, POSIXlt, Date, Period, chron, yearmon, yearqtr,
zoo, zooreg, timeDate, xts, its, ti, jul, timeSeries, and fts
objects.

\code{isoyear()} returns years according to the ISO 8601 week calendar.

\code{epiyear()} returns years according to the epidemilogical week calendars.
}
\details{
year does not yet support years before 0 C.E.
}
\examples{
get_current_year()
}
\references{
\url{http://en.wikipedia.org/wiki/ISO_week_date}
\url{http://www.cmmcp.org/epiweek.htm}
}
