% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_paths.R
\name{get_project_paths}
\alias{get_project_paths}
\title{Create paths to an analysis project}
\usage{
get_project_paths(project, parent = NULL)
}
\arguments{
\item{project}{String giving the name of a project directory.}

\item{parent}{String giving the path or the parent directory of the
\code{project}.}
}
\value{
A character vector.
}
\description{
This function creates paths that represent the structure of a common
"analysis project".
}
\examples{
get_project_paths("a-project")

# Use `fs::dir_crete() to create all directories in one step
project_paths <- get_project_paths("a-project", parent = tempdir())
fs::dir_create(project_paths)

all(fs::dir_exists(project_paths))

# Cleanup
fs::dir_delete(project_paths)
}
\seealso{
\code{\link[fs:dir_create]{fs::dir_create()}}
}
