% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda_calculation.R
\name{get_ref_sector}
\alias{get_ref_sector}
\title{Default value for the \code{ref_sector} argument to \code{\link[=sda_calculation]{sda_calculation()}}}
\usage{
get_ref_sector()
}
\value{
A character vector.
}
\description{
Default value for the \code{ref_sector} argument to \code{\link[=sda_calculation]{sda_calculation()}}
}
\examples{
get_ref_sector()
}
