% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda_portfolio_target.R
\name{get_sectors}
\alias{get_sectors}
\title{Default value for the \code{ref_sector} argument to \code{\link[=sda_portfolio_target]{sda_portfolio_target()}}}
\usage{
get_sectors()
}
\value{
A character vector.
}
\description{
Default value for the \code{ref_sector} argument to \code{\link[=sda_portfolio_target]{sda_portfolio_target()}}
}
\examples{
get_sectors()
}
