% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_indicator.R
\name{influencemap_weighting_methodology}
\alias{influencemap_weighting_methodology}
\title{TODO \@vintented}
\usage{
influencemap_weighting_methodology(
  input_results,
  input_audit,
  metric_name = "temperature",
  group_vars = c("Investor.Name", "Portfolio.Name"),
  sector_weightings
)
}
\arguments{
\item{sector_weightings}{}
}
\value{

}
\description{
TODO \@vintented
}
