% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_input_files.R
\name{input_file_names}
\alias{input_file_names}
\title{Name of common input files}
\usage{
input_file_names()
}
\value{
A character vector.
}
\description{
Name of common input files
}
\examples{
input_file_names()
}
