% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_portfolio_ald_scenario.R
\name{join_portfolio_ald_scenario}
\alias{join_portfolio_ald_scenario}
\title{Join a matched-loanbook object to the ald and scenario}
\usage{
join_portfolio_ald_scenario(match_result, ald, scenario)
}
\arguments{
\item{match_result}{A dataframe like the output of
\code{\link[r2dii.match:prioritize]{r2dii.match::prioritize()}}.}

\item{ald}{An asset level dataframe like \link[r2dii.dataraw:ald_demo]{r2dii.dataraw::ald_demo}.}

\item{scenario}{A scenario dataframe like \link[r2dii.dataraw:scenario_demo]{r2dii.dataraw::scenario_demo}.}
}
\value{
Returns a fully joined dataframe, linking portfolio, ald and
scenario.
}
\description{
\code{join_portfolio_ald_scenario()} is a simple wrapper of several calls to
\code{join()} functions, forming the master dataset to be used in later steps of
the analysis.
}
\examples{
installed <- requireNamespace("r2dii.dataraw", quietly = TRUE) &&
  requireNamespace("r2dii.match", quietly = TRUE)
if (installed) {
  library(r2dii.dataraw)
  library(r2dii.match)

  # Pretending you validated this dataset
  validated_match <- prioritize(match_name(loanbook_demo, ald_demo))

  validated_match \%>\%
    join_portfolio_ald_scenario(
      ald = ald_demo, scenario = add_fair_share_columns(scenario_demo, 2020)
    )
}
}
