% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_indicator.R
\name{map_sector_exposure}
\alias{map_sector_exposure}
\title{TODO vintented}
\usage{
map_sector_exposure(input_audit)
}
\arguments{
\item{input_audit}{A dataframe like \link[r2dii.analysis:sample_audit]{r2dii.analysis::sample_audit}.}
}
\value{
Returns a dataframe with the value_usd invested in each sector within a portfolio.
}
\description{
TODO vintented
}
