% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_indicator.R
\name{mapped_sector_exposure}
\alias{mapped_sector_exposure}
\title{TODO \@vintented}
\usage{
mapped_sector_exposure(input_audit)
}
\arguments{
\item{input_audit}{}
}
\value{

}
\description{
TODO \@vintented
}
