% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_project_files.R
\name{path_input_files}
\alias{path_input_files}
\title{Create paths to common input files}
\usage{

  path_input_files(path_dir = r2dii.utils::path_dropbox_2dii("Portcheck_v2",
  "00_Administration", "20_Input_Files"),
  file_names = c("ProjectName_Input.csv", "ReportParameterFile.yml",
  "AnalysisParameters.yml"))
}
\arguments{
\item{path_dir}{String. Path to the directory containing \code{file_names}}

\item{file_names}{Character vector giving the name of the files in
\code{path_dir}.}
}
\value{
A character vector.
}
\description{
Create paths to common input files
}
\examples{
path_input_files()

# The argument defaults should be okay but you can change them if needed
# This is useful if the name of a directory changes
path_input_files(file_names = c("new_file_1", "new_file_2"))
# This is useful for tests and examples
path_input_files(path_dir = tempdir())
}
