% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_project.R
\name{path_project}
\alias{path_project}
\title{Create paths in 2dii's dropbox under "PortCheck_v2/10_Projects"}
\usage{
path_project(project, ..., parent = NULL)
}
\arguments{
\item{project}{String giving the name of a project directory.}

\item{...}{character vectors, if any values are NA, the result will also be
NA.}

\item{parent}{String giving the path or the parent directory of the
\code{project}.}
}
\value{
A "character", subclass "fs_path".
}
\description{
This function output a path inside 2dii's dropbox folder, under
"PortCheck_v2/10_Projects/<a directory>", where "<a directory>" is usually a
directory within which you may need to access multiple subdirectories. This
function replaces the legacy global variable \code{project_location}.
}
\examples{
path_project("a-project")

path_project("a-project", "a", "sub", "directory")

# Useful for tests and examples
path_project("a-project", "a", "sub", "directory", parent = tempdir())

# Common paths for analyses --------------------------------------------
path_project("a-project", "00_Log_Files")
path_project("a-project", "10_Parameter_File")
path_project("a-project", "20_Raw_Inputs")
path_project("a-project", "30_Processed_Inputs")
path_project("a-project", "40_Results")
path_project("a-project", "50_Outputs")
}
