% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_project_dirs.R
\name{path_project_dirs}
\alias{path_project_dirs}
\title{Create paths to common project directories}
\usage{
path_project_dirs(project, parent = NULL)
}
\arguments{
\item{project}{String giving the name of a project directory.}

\item{parent}{String giving the path or the parent directory of the
\code{project}.}
}
\value{
A character vector.
}
\description{
This function creates paths that represent the structure of a common
"analysis project".
}
\examples{
path_project_dirs("a-project")

# Use `fs::dir_crete() to create all directories in one step
path_to_project_dirs <- path_project_dirs("a-project", parent = tempdir())
fs::dir_create(path_to_project_dirs)

all(fs::dir_exists(path_to_project_dirs))

# Cleanup
fs::dir_delete(path_to_project_dirs)
}
\seealso{
\code{\link[fs:dir_create]{fs::dir_create()}}
}
