% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_input_check.R
\name{portfolio_input_check}
\alias{portfolio_input_check}
\title{Title}
\usage{
portfolio_input_check(project)
}
\arguments{
\item{project}{String giving the name of a project directory.}
}
\description{
Title
}
