% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2dii.analysis-package.R
\docType{package}
\name{r2dii.analysis-package}
\alias{r2dii.analysis}
\alias{r2dii.analysis-package}
\title{r2dii.analysis: Measure Climate Scenario Alignment of Corporate Loans}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

These tools help you to assess if a corporate lending portfolio aligns with climate goals. They summarize key climate indicators attributed to the portfolio (e.g. production, emission factors), and calculate alignment targets based on climate scenarios. They implement in R the last step of the free software 'PACTA' (Paris Agreement Capital Transition Assessment; \url{https://www.transitionmonitor.com/}). Financial institutions use 'PACTA' to study how their capital allocation decisions align with climate change mitigation goals.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RMI-PACTA/r2dii.analysis}
  \item \url{https://rmi-pacta.github.io/r2dii.analysis/}
  \item Report bugs at \url{https://github.com/RMI-PACTA/r2dii.analysis/issues}
}

}
\author{
\strong{Maintainer}: Jackson Hoffart \email{jackson.hoffart@gmail.com} (\href{https://orcid.org/0000-0002-8600-5042}{ORCID})

Authors:
\itemize{
  \item Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID}) [contractor]
  \item Klaus Hogedorn \email{klaus@2degrees-investing.org}
}

Other contributors:
\itemize{
  \item Rocky Mountain Institute \email{PACTA4banks@rmi.org} [copyright holder, funder]
}

}
\keyword{internal}
