% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-single_indicator.R
\docType{data}
\name{sample_results}
\alias{sample_results}
\title{Sample daset for \code{single_indicator()} and friends}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 714 rows and 13 columns.}
\usage{
sample_results
}
\description{
TODO \@vintented please document
}
\keyword{datasets}
