% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-single_indicator.R
\docType{data}
\name{scenario_relationships}
\alias{scenario_relationships}
\title{Sample daset for \code{single_indicator()} and friends}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12 rows and 4 columns.}
\usage{
scenario_relationships
}
\description{
TODO \@vintented please document
}
\keyword{datasets}
