% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda_calculation.R
\name{sda_calculation}
\alias{sda_calculation}
\title{TODO \@vintented}
\usage{
sda_calculation(
  market_data,
  port_data,
  ref_sector = get_ref_sector(),
  ref_scenario = "B2DS",
  ref_geography = "Global",
  start_year = r2dii.utils::START.YEAR(),
  target_year = 2040
)
}
\arguments{
\item{market_data}{A dataframe with market results.}

\item{port_data}{A dataframe with portfolio results data.}

\item{ref_sector}{A character vector giving one or more sectors (with
emissions factors) to calculate the SDA.}

\item{ref_scenario}{A character vector giving one or more scenarios to use as
the SDA target.}

\item{ref_geography}{A character vector giving one or more scenario
geographies for each scenario.}

\item{start_year}{A numeric or character vector giving the start year used in
the SDA calculation as a numeric (i.e. start_year == 2020).}

\item{target_year}{A numeric or character vector giving the end year used in
the SDA calculation as a numeric (i.e. target_year == 2045). By default the
function will use the last year in the current market data.}
}
\value{
Returns a dataframe where the \code{Scen.Sec.EmissionsFactor} column
holds the result of the SDA calculation.
}
\description{
TODO \@vintented
}
\examples{
# Use example configuration-files
library(r2dii.utils)

# Use a toy configuration file
restore_options <- options(r2dii_config = example_config("config_demo.yml"))
on.exit(restore_options)

# Use `start_year` from the configuration file
START.YEAR()
sda_calculation(market, portfolio)

sda_calculation(market, portfolio, ref_sector = "Steel")

# This configuration file lacks `start_year`
options(r2dii_config = example_config("config-toy.yml"))
START.YEAR()

# Fails
try(sda_calculation(market, portfolio))

# Passes
sda_calculation(market, portfolio, start_year = "2019")
}
\seealso{
\code{\link[r2dii.utils:get_config]{r2dii.utils::get_config()}}, \code{\link[r2dii.utils:START.YEAR]{r2dii.utils::START.YEAR()}}
}
