% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda_portfolio_target.R
\name{sda_portfolio_target}
\alias{sda_portfolio_target}
\title{Calculate portfolio targets using the Sectoral Decarbonisation Approach (SDA)}
\usage{
sda_portfolio_target(
  market,
  portfolio,
  ref_scenario = "B2DS",
  ref_geography = "Global",
  ref_sector = NULL,
  start_year = NULL,
  target_year = NULL
)
}
\arguments{
\item{market}{A dataframe like \link[r2dii.analysis:market]{r2dii.analysis::market}.}

\item{portfolio}{A dataframe like \link[r2dii.analysis:portfolio]{r2dii.analysis::portfolio}.}

\item{ref_scenario}{A character vector giving one or more scenarios to use as
the SDA target.}

\item{ref_geography}{A character vector giving one or more scenario
geographies for each scenario.}

\item{ref_sector}{A character vector giving one or more sectors present in
both \code{market} and \code{portfolio} data. \code{NULL} defaults to all expected sectors
(see section See Also).}

\item{start_year}{A length-1 numeric or character vector giving the start
year used in the SDA calculation. \code{NULL} defaults to extracting the year
from a configuration file (see section See Also).}

\item{target_year}{A length-1 numeric or character vector giving the end year
used in the SDA calculation. It must be a year present in all \code{ref_sectors}
of \code{market}. \code{NULL} defaults to the latest year shared across all sectors
given by \code{ref_sector} and found in \code{market}.}
}
\value{
Returns a dataframe where the \code{Scen.Sec.EmissionsFactor} column
holds the result of the SDA calculation.
}
\description{
The Sectoral Decarbonisation Approach (SDA) is a method for setting corporate
emission reduction targets in line with climate science.
}
\examples{
# Use example configuration-files
library(r2dii.utils)

# Use a toy configuration file
restore_options <- options(r2dii_config = example_config("config_demo.yml"))
on.exit(restore_options)

# Use `start_year` from the configuration file
START.YEAR()
sda_portfolio_target(market, portfolio)

sda_portfolio_target(market, portfolio, ref_sector = "Steel")

# This configuration file lacks `start_year`
options(r2dii_config = example_config("config-toy.yml"))
START.YEAR()

# Fails
try(sda_portfolio_target(market, portfolio))

# Passes
sda_portfolio_target(market, portfolio, start_year = "2019")
}
\seealso{
\code{\link[r2dii.utils:get_config]{r2dii.utils::get_config()}}, \code{\link[r2dii.utils:START.YEAR]{r2dii.utils::START.YEAR()}},
\code{\link[=sectors]{sectors()}}.
}
