% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_indicator.R
\name{single_indicator}
\alias{single_indicator}
\title{TODO \@vintented}
\usage{
single_indicator(
  input_results,
  upper_temp_threshold = 6,
  lower_temp_threshold = 1.5,
  start_year = 2019,
  time_horizon = 5,
  allocation = "PortfolioWeight",
  production_type = "absolute",
  group_vars = c("Investor.Name", "Portfolio.Name", "Asset.Type"),
  scenario_relationships
)
}
\arguments{
\item{scenario_relationships}{}
}
\value{

}
\description{
TODO \@vintented
}
