% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_company_production.R
\name{summarize_company_production}
\alias{summarize_company_production}
\title{Summarize the portfolio-weighted produciton at company level}
\usage{
summarize_company_production(data, ..., use_credit_limit = FALSE)
}
\arguments{
\item{data}{A "data.frame" like the output of \code{\link[=join_ald_scenario]{join_ald_scenario()}}.}

\item{...}{Variables to group by.}

\item{use_credit_limit}{Logical vector of length 1. \code{FALSE} defaults to
using the column \code{loan_size_outstanding}. Set to \code{TRUE} to use the column
\code{loan_size_credit_limit} instead.}
}
\value{
A tibble with the same groups as the input (if any) and columns:
\code{sector}, \code{technology}, \code{year}, and \code{weighted_production}.
\code{weighted_production} results are aggregated to the company-level.
}
\description{
A simpler wrapper used to calculate the loan-weighted production for each
matched company, and aggregate to the company level.
}
\examples{
library(r2dii.data)

master <- r2dii.data::loanbook_demo \%>\%
  r2dii.match::match_name(r2dii.data::ald_demo) \%>\%
  r2dii.match::prioritize() \%>\%
  join_ald_scenario(r2dii.data::ald_demo, r2dii.scenario::scenario_demo)

summarize_company_production(master)

summarize_company_production(master, use_credit_limit = TRUE)
}
