% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_portfolio_production.R
\name{summarize_portfolio_production}
\alias{summarize_portfolio_production}
\title{Summarize the portfolio-weighted production at portfolio level}
\usage{
summarize_portfolio_production(data, ..., use_credit_limit = FALSE)
}
\arguments{
\item{data}{A "data.frame" like the output of \code{\link[=join_ald_scenario]{join_ald_scenario()}}.}

\item{...}{
  Arguments passed on to \code{\link[=summarize_weighted_production]{summarize_weighted_production}}
  \describe{
    \item{\code{}}{}
  }}

\item{use_credit_limit}{Logical vector of length 1. \code{FALSE} defaults to using
the column \code{loan_size_outstanding}. Set to \code{TRUE} to use the column
\code{loan_size_credit_limit} instead.}
}
\value{
A tibble with the same groups as the input (if any) and columns:
\code{sector}, \code{technology}, \code{year}, and \code{weighted_production}.
}
\description{
A simpler wrapper used to calculate the loan-weighted production for each
matched company, and aggregate to the portfolio level.
}
\examples{
master <- tibble::tibble(
  id_loan = c("i1", "i2", "i1", "i2"),
  loan_size_outstanding = c(40, 10, 40, 10),
  loan_size_credit_limit = c(2, 2, 2, 2),
  sector = c("automotive", "automotive","automotive","automotive"),
  name_ald = c("shaanxi auto","shaanxi auto","shaanxi auto", "shaanxi auto"),
  technology = c("ta", "ta", "tb", "tb"),
  year = c(2025, 2025, 2025, 2025),
  production = c(10, 30, 20, 40),
  plant_location = c("BF", "BF", "BF", "BF"),
  scenario = c("sds", "sds", "sds", "sds"),
  region = c("global", "global", "global", "global"),
  tmsr = c(0.5, 0.5, 0.5, 0.5),
  smsp = c(-0.08, -0.08, -0.08, -0.08)
)
master

summarize_portfolio_production(master)

summarize_portfolio_production(master, use_credit_limit = TRUE)
}
\seealso{
\code{\link[=join_ald_scenario]{join_ald_scenario()}}.
}
