% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_fair_share_company.R
\name{target_fair_share_company}
\alias{target_fair_share_company}
\title{Add company-level production targets}
\usage{
target_fair_share_company(data)
}
\arguments{
\item{data}{A "data.frame" like the output of
\code{\link[=summarize_company_production]{summarize_company_production()}}.}
}
\value{
A tibble with the same groups as the input (if any) and columns:
\code{tmsr_target_weighted_production} and \code{smsp_target_weighted_production}.
}
\description{
This function calculates the company-level production targets, as
calculated using either the technology market share ratio \code{tmsr} or sector
market share percentage \code{smsp}.
}
\examples{
library(r2dii.analysis)
library(r2dii.data)
library(r2dii.match)

master <- r2dii.data::loanbook_demo \%>\%
  r2dii.match::match_name(r2dii.data::ald_demo) \%>\%
  r2dii.match::prioritize() \%>\%
  join_ald_scenario(r2dii.data::ald_demo,
    r2dii.data::scenario_demo_2020,
    region_isos = region_isos_demo
  )

company_production <- summarize_company_production(master)
company_production

target_fair_share_company(company_production)
}
