% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_market_share_portfolio.R
\name{target_market_share_portfolio}
\alias{target_market_share_portfolio}
\title{Add portfolio-level production targets}
\usage{
target_market_share_portfolio(data)
}
\arguments{
\item{data}{A "data.frame" like the output of
\code{\link[=summarize_portfolio_production]{summarize_portfolio_production()}}.}
}
\value{
A tibble with the same groups as the input (if any) and columns:
\code{tmsr_target_weighted_production} and \code{smsp_target_weighted_production}.
}
\description{
This function calculates the portfolio-level production targets, as
calculated using either the technology market share ratio \code{tmsr} or sector
market share percentage \code{smsp}.
}
\examples{
library(r2dii.analysis)
library(r2dii.data)
library(r2dii.match)

master <- r2dii.data::loanbook_demo \%>\%
  r2dii.match::match_name(r2dii.data::ald_demo) \%>\%
  r2dii.match::prioritize() \%>\%
  join_ald_scenario(r2dii.data::ald_demo,
    r2dii.data::scenario_demo_2020,
    region_isos = region_isos_demo
  )

portfolio_production <- summarize_portfolio_production(master)
portfolio_production

target_market_share_portfolio(portfolio_production)
}
\seealso{
Other functions to calculate scenario targets: 
\code{\link{target_market_share_company}()},
\code{\link{target_sda}()}
}
\concept{functions to calculate scenario targets}
