% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-single_indicator.R
\docType{data}
\name{tech_sector_weighting}
\alias{tech_sector_weighting}
\title{Sample daset for \code{single_indicator()} and friends}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 33 rows and 4 columns.}
\usage{
tech_sector_weighting
}
\description{
Sample daset for \code{single_indicator()} and friends
}
\keyword{datasets}
