% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_company_target.R
\name{add_company_target}
\alias{add_company_target}
\title{Add company-level targets}
\usage{
add_company_target(data)
}
\arguments{
\item{data}{A "data.frame" like the output of
\code{\link[=summarize_company_production]{summarize_company_production()}}.}
}
\value{
A tibble with the same groups as the input (if any) and columns:
\code{tmsr_target_weighted_production} and \code{smsp_target_weighted_production}.
}
\description{
This function calculates the company-level production targets, as
calculated using either the technology market share ratio \code{tmsr} or sector
market share percentage \code{smsp}.
}
\examples{
master <- tibble::tibble(
  id_loan = c("i1", "i2", "i1", "i2"),
  loan_size_outstanding = c(40, 10, 40, 10),
  loan_size_credit_limit = c(2, 2, 2, 2),
  sector = c("automotive", "automotive","automotive","automotive"),
  name_ald = c("shaanxi auto","shaanxi auto","shaanxi auto", "shaanxi auto"),
  technology = c("ta", "ta", "tb", "tb"),
  year = c(2025, 2025, 2025, 2025),
  production = c(10, 30, 20, 40),
  plant_location = c("BF", "BF", "BF", "BF"),
  scenario = c("sds", "sds", "sds", "sds"),
  region = c("global", "global", "global", "global"),
  tmsr = c(0.5, 0.5, 0.5, 0.5),
  smsp = c(-0.08, -0.08, -0.08, -0.08)
)
master

company_production <- summarize_company_production(master)
company_production

add_company_target(company_production)
}
