% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_company_target.R
\name{add_company_target}
\alias{add_company_target}
\title{Add company-level targets}
\usage{
add_company_target(data)
}
\arguments{
\item{data}{A "data.frame" like the output of
\code{\link[=summarize_company_production]{summarize_company_production()}}.}
}
\value{
A tibble with the same groups as the input (if any) and columns:
\code{tmsr_target_weighted_production} and \code{smsp_target_weighted_production}.
}
\description{
This function calculates the company-level production targets, as
calculated using either the technology market share ratio \code{tmsr} or sector
market share percentage \code{smsp}.
}
\examples{
library(r2dii.analysis)
library(r2dii.data)
library(r2dii.match)

master <- r2dii.data::loanbook_demo \%>\%
  r2dii.match::match_name(r2dii.data::ald_demo) \%>\%
  r2dii.match::prioritize() \%>\%
  join_ald_scenario(r2dii.data::ald_demo, r2dii.data::scenario_demo_2020)

company_production <- summarize_company_production(master)
company_production

add_company_target(company_production)
}
