% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_portfolio_target.R
\name{add_portfolio_target}
\alias{add_portfolio_target}
\title{Add portfolio-level targets}
\usage{
add_portfolio_target(data)
}
\arguments{
\item{data}{A "data.frame" like the output of
\code{\link[=summarize_portfolio_production]{summarize_portfolio_production()}}.}
}
\value{
A tibble with the same groups as the input (if any) and columns:
\code{tmsr_target_weighted_production} and \code{smsp_target_weighted_production}.
}
\description{
This function calculates the portfolio-level production targets, as
calculated using either the technology market share ratio \code{tmsr} or sector
market share percentage \code{smsp}.
}
\examples{
installed <- requireNamespace("r2dii.data", quietly = TRUE) &&
  requireNamespace("r2dii.match", quietly = TRUE)
if (!installed) stop("Please install r2dii.match and r2dii.data")

library(r2dii.data)
library(r2dii.match)

master <- loanbook_demo \%>\%
  match_name(ald_demo) \%>\%
  prioritize() \%>\%
  join_ald_scenario(ald_demo, scenario_demo_2020)

portfolio_production <- summarize_portfolio_production(master, tmsr, smsp)

add_portfolio_target(portfolio_production)
}
