% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_indicator.R
\name{apply_influencemap_portfolio_weighting}
\alias{apply_influencemap_portfolio_weighting}
\title{TODO vintented}
\usage{
apply_influencemap_portfolio_weighting(
  input_results,
  input_audit,
  metric_name = "temperature",
  group_vars = c("Investor.Name", "Portfolio.Name", "Asset.Type")
)
}
\arguments{
\item{sector_weightings}{}
}
\value{

}
\description{
TODO vintented
}
