% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_indicator.R
\name{calculate_temperature_indicator}
\alias{calculate_temperature_indicator}
\title{TODO Add title}
\usage{
calculate_temperature_indicator(
  input_results,
  upper_temp_threshold = 6,
  lower_temp_threshold = 1.5,
  start_year = NULL,
  time_horizon = NULL,
  allocation = NULL,
  production_type = "absolute",
  group_vars = c("Investor.Name", "Portfolio.Name", "Asset.Type")
)
}
\arguments{
\item{input_results}{A dataframe like \link[r2dii.analysis:sample_results]{r2dii.analysis::sample_results}.}

\item{upper_temp_threshold}{A numeric value that defines the upper temperature
threshold to cut outlier values.}

\item{lower_temp_threshold}{A numeric value that defines the lower temperature
threshold to cut outlier values.}

\item{start_year}{A numeric value that defines the start year of the analysis.}

\item{time_horizon}{A numeric values that defines the time horizon of the analysis.}

\item{allocation}{A character value or list of characters that define the portfolio approach to use.}

\item{production_type}{A character value (absolute or reletive) that defines
whether to use the stock or flow production for a technology.}

\item{group_vars}{A set of grouping variables that define the granularity of the metric.}
}
\value{

}
\description{
TODO Add title
}
