% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2_intensity_scenario.R
\docType{data}
\name{co2_intensity_scenario}
\alias{co2_intensity_scenario}
\title{A prepared co2 intensity climate scenario dataset for demonstration}
\format{
\code{co2_intensity_scenario} is a \link{data.frame} with columns:
\itemize{
\item \code{region} (character): The region to which the pathway is relevant.
\item \code{scenario} (character): The name of the scenario.
\item \code{scenario_source} (character): The source publication from which the
scenario was taken.
\item \code{sector} (character): The sector to which the scenario prescribes a
pathway.
\item \code{smsp} (double): Sector market share percentage of the pathway calculated
in 2020.
\item \code{technology} (character): The technology within the sector to which the
scenario prescribes a pathway.
\item \code{tmsr} (double): Technology market share ratio of the pathway calculated in
}
\enumerate{
\item 
}
\itemize{
\item \code{year} (integer): The year at which the pathway value is prescribed.
}
}
\usage{
co2_intensity_scenario
}
\description{
Fake co2 intensity climate scenario dataset, prepared for the software PACTA (Paris
Agreement Capital Transition Assessment). It imitates climate scenario data
(e.g. from the International Energy Agency (IEA)) including the change
through time in production across industrial sectors (calculated by
\href{https://2degrees-investing.org/}{2DII}).
}
\examples{
head(co2_intensity_scenario)
}
\seealso{
\link{data_dictionary}

Other demo datasets: 
\code{\link{market}},
\code{\link{portfolio}}
}
\concept{demo datasets}
\keyword{datasets}
