% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_ald_scenario.R
\name{join_ald_scenario}
\alias{join_ald_scenario}
\title{Join a data-loanbook object to the ald and scenario}
\usage{
join_ald_scenario(data, ald, scenario)
}
\arguments{
\item{data}{A dataframe like the output of
\code{\link[r2dii.match:prioritize]{r2dii.match::prioritize()}}.}

\item{ald}{An asset level dataframe like \link[r2dii.dataraw:ald_demo]{r2dii.dataraw::ald_demo}.}

\item{scenario}{A scenario dataframe like \link[r2dii.dataraw:scenario_demo]{r2dii.dataraw::scenario_demo}.}
}
\value{
Returns a fully joined dataframe, linking portfolio, ald and
scenario.
}
\description{
\code{join_ald_scenario()} is a simple wrapper of several calls to
\code{join()} functions, forming the master dataset to be used in later steps of
the analysis.
}
\examples{
installed <- requireNamespace("r2dii.dataraw", quietly = TRUE) &&
  requireNamespace("r2dii.match", quietly = TRUE)
if (!installed) stop("Please install r2dii.match and r2dii.dataraw")

library(r2dii.dataraw)
library(r2dii.match)

# Example of valid matches after `r2dii.match::match_name()` and manual edits
path <- system.file("extdata", "valid_matches.csv", package = "r2dii.analysis")
valid_matches <- tibble::as_tibble(read.csv(path, stringsAsFactors = FALSE))

valid_matches \%>\%
  join_ald_scenario(
    ald = ald_demo,
    scenario = scenario_demo
  )
}
