% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_ald_scenario.R
\name{join_ald_scenario}
\alias{join_ald_scenario}
\title{Join a data-loanbook object to the ald and scenario}
\usage{
join_ald_scenario(
  data,
  ald,
  scenario,
  region_isos = r2dii.data::region_isos,
  add_green_technologies = FALSE
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated in favour of \code{\link[=join_abcd_scenario]{join_abcd_scenario()}}. See \href{https://2degreesinvesting.github.io/posts/2022-03-02-ald-becomes-abcd/}{here}
for more information.
}
\examples{
\dontrun{
installed <- requireNamespace("r2dii.data", quietly = TRUE) &&
  requireNamespace("r2dii.match", quietly = TRUE) &&
  packageVersion("r2dii.match") >= "0.1.0"

if (installed) {
  library(r2dii.data)
  library(r2dii.match)

  valid_matches <- match_name(loanbook_demo, ald_demo) \%>\%
    prioritize()

  valid_matches \%>\%
    join_ald_scenario(
      ald = ald_demo,
      scenario = scenario_demo_2020,
      region_isos = region_isos_demo
    )

   # ->

  valid_matches \%>\%
    join_abcd_scenario(
      abcd = abcd_demo,
      scenario = scenario_demo_2020,
      region_isos = region_isos_demo
    )
}
}
}
\keyword{internal}
