% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_input_files.R
\name{path_input_files}
\alias{path_input_files}
\alias{path_output_files}
\title{Create paths to common input and output files}
\usage{
path_input_files(parent = r2dii.utils::path_dropbox_2dii("Portcheck_v2",
  "00_Administration", "20_Input_Files"),
  file_names = input_file_names())

path_output_files(project, parent = NULL, file_names = NULL)
}
\arguments{
\item{parent}{String. Path to the directory containing \code{file_names}}

\item{file_names}{Character vector giving the name of the files in
\code{parent}.}

\item{project}{String giving the name of a project directory.}
}
\value{
A character vector.
}
\description{
Create paths to common input and output files
}
\examples{
path_input_files()

path_output_files("a-project")

# Use fs::file_copy() to copy all files in one step
\dontrun{
fs::file_copy(
  path_input_files(),
  path_output_files("a-project"),
  overwrite = TRUE
)
}

# Use `file_names` to create pahts to new files
path_input_files(file_names = c("new_file_1", "new_file_2"))
path_output_files("a-project", file_names = c("new_file_1", "new_file_2"))

# Use `parent` to change the parent directory
path_input_files(parent = tempdir())
path_output_files("a-project", parent = tempdir())

}
