% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.R
\docType{data}
\name{portfolio}
\alias{portfolio}
\title{TODO \@vintented: Add title}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 24 rows and 9 columns.}
\source{
\@vintented via <https://github.com/2DegreesInvesting/
r2dii.analysis/issues/6#issuecomment-561543097>.
}
\usage{
portfolio
}
\description{
TODO \@vintented: Add description
}
\keyword{datasets}
