% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_portfolio_file.R
\name{read_raw_portfolio}
\alias{read_raw_portfolio}
\title{Read raw portfolio data}
\usage{
read_raw_portfolio(project)
}
\arguments{
\item{project}{String giving the name of a project directory.}
}
\value{
A \link[tibble:tibble]{tibble::tibble}.
}
\description{
Read raw portfolio data
}
\examples{
if (r2dii.utils::dropbox_exists()) {
  read_raw_portfolio("TEST")
}
}
