% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_raw_portfolio_file.R
\name{read_raw_portfolio}
\alias{read_raw_portfolio}
\title{Read raw portfolio data}
\usage{
read_raw_portfolio(path, delim = ",", ...)
}
\arguments{
\item{path}{A character string giving the path to a raw portfolio file.}

\item{delim}{Passed to \link[readr:read_delim]{readr::read_delim}.}

\item{...}{Arguments passed to \link[readr:read_delim]{readr::read_delim}.}
}
\value{
A \link[tibble:tibble]{tibble::tibble}.
}
\description{
Read raw portfolio data
}
\examples{
"raw_portfolio.csv" \%>\%
  path_example("r2dii.analysis") \%>\%
  read_raw_portfolio()
}
\seealso{
\code{\link[=find_project_input_files]{find_project_input_files()}}
}
