% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-single_indicator.R
\docType{data}
\name{sample_audit}
\alias{sample_audit}
\title{Sample daset for \code{single_indicator()} and friends}
\format{An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 18 rows and 5 columns.}
\usage{
sample_audit
}
\description{
TODO vintented please document
}
\keyword{datasets}
