% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda_calculation.R
\name{sda_calculation}
\alias{sda_calculation}
\title{TODO \@vintented}
\usage{
sda_calculation(
  market_data,
  port_data,
  ref_sector = c("Cement", "Steel"),
  ref_scenario = "B2DS",
  ref_geography = "Global",
  start_year = get_current_year(),
  target_year = 2040
)
}
\arguments{
\item{market_data}{TODO \@vintented}

\item{port_data}{TODO \@vintented}

\item{ref_sector}{TODO \@vintented}

\item{ref_scenario}{TODO \@vintented}

\item{ref_geography}{TODO \@vintented}

\item{start_year}{TODO \@vintented}

\item{target_year}{TODO \@vintented}
}
\value{
TODO \@vintented
}
\description{
TODO \@vintented
}
\examples{
sda_calculation(
  r2dii.analysis::market,
  r2dii.analysis::portfolio
)
}
