% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda_calculation.R
\name{sda_calculation}
\alias{sda_calculation}
\title{TODO \@vintented}
\usage{
sda_calculation(
  market_data,
  port_data,
  ref_sector = get_ref_sector(),
  ref_scenario = "B2DS",
  ref_geography = "Global",
  start_year = get_current_year(),
  target_year = 2040
)
}
\arguments{
\item{market_data}{A dataframe with market results.}

\item{port_data}{A dataframe with portfolio results data.}

\item{ref_sector}{A list of one or more sectors (with emissions factors) to
calculate the SDA.}

\item{ref_scenario}{A list of one or more scenarios to use as the SDA target.}

\item{ref_geography}{A list of one or more scenario geographies for each
scenario.}

\item{start_year}{The start year used in the SDA calculation as a numeric
(i.e. start_year == 2020). By default the function will use the config.
start year.}

\item{target_year}{The end year used in the SDA calculation as a numeric
(i.e. target_year == 2045). By default the function will use the last year
in the current market data.}
}
\value{
The SDA function returns the portfolio results dataframe with the
"scenario sector emissions factor" column overwritten with the portfolio
SDA calculation.
}
\description{
TODO \@vintented
}
\examples{
sda_calculation(
  r2dii.analysis::market,
  r2dii.analysis::portfolio
)
}
