% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_portfolio_production.R
\name{summarize_portfolio_production}
\alias{summarize_portfolio_production}
\title{Summarize the portfolio-weighted production at portfolio level}
\usage{
summarize_portfolio_production(data, ..., use_credit_limit = FALSE)
}
\arguments{
\item{data}{A "data.frame" like the output of \code{\link[=join_ald_scenario]{join_ald_scenario()}}.}

\item{...}{
  Arguments passed on to \code{\link[=summarize_weighted_production]{summarize_weighted_production}}
  \describe{
    \item{\code{}}{}
  }}

\item{use_credit_limit}{Logical vector of length 1. \code{FALSE} defaults to using
the column \code{loan_size_outstanding}. Set to \code{TRUE} to use the column
\code{loan_size_credit_limit} instead.}
}
\value{
A tibble with the same groups as the input (if any) and columns:
\code{sector}, \code{technology}, \code{year}, and \code{weighted_production}.
}
\description{
A simpler wrapper used to calculate the loan-weighted production for each
matched company, and aggregate to the portfolio level.
}
\examples{
installed <- requireNamespace("r2dii.data", quietly = TRUE) &&
  requireNamespace("r2dii.match", quietly = TRUE)
if (!installed) stop("Please install r2dii.match and r2dii.data")

library(r2dii.data)
library(r2dii.match)

master <- loanbook_demo \%>\%
  match_name(ald_demo) \%>\%
  prioritize() \%>\%
  join_ald_scenario(ald_demo, scenario_demo_2020)

summarize_portfolio_production(master)

summarize_portfolio_production(master, use_credit_limit = TRUE)
}
\seealso{
\code{\link[=join_ald_scenario]{join_ald_scenario()}}.
}
